+++ tests/test_structure.py
@@ -6,7 +6,7 @@ from typing import NewType
 import pytest
 from compages import (
     DataclassBase,
-    StructureDictIntoDataclass,
+    StructureMappingIntoDataclass,
     Structurer,
     StructuringError,
     simple_structure,
@@ -71,7 +71,7 @@ def test_structure_routing():
             HexInt: structure_hex_int,
             list[int]: structure_custom_generic,
             list: structure_into_list,
-            DataclassBase: StructureDictIntoDataclass(),
+            DataclassBase: StructureMappingIntoDataclass(),
         }
     )
 
@@ -101,7 +101,7 @@ def test_structure_generators():
         {
             int: structure_into_int,
             Container: structure_container,
-            DataclassBase: StructureDictIntoDataclass(),
+            DataclassBase: StructureMappingIntoDataclass(),
         }
     )
 
@@ -157,7 +157,7 @@ def test_error_rendering():
             dict: structure_into_dict,
             int: structure_into_int,
             str: structure_into_str,
-            DataclassBase: StructureDictIntoDataclass(),
+            DataclassBase: StructureMappingIntoDataclass(),
         }
     )